pyproject.toml
[tool.poetry]
name = "market-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.114.2"
sqlalchemy = "^2.0.34"
loguru = "^0.7.2"
alembic = "^1.13.2"
uvicorn = "^0.30.6"
pydantic-settings = "^2.5.2"
asyncpg = "^0.29.0"
greenlet = "^3.1.0"
aiohttp = "^3.10.5"
pytoniq = "^0.1.39"
aiogram = "^3.13.1"
redis = "^5.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
